+++ src/nycti/discord/registration.py
@@ -10,7 +10,6 @@ from nycti.discord.logs import register_logs_command
 from nycti.discord.memory import register_memory_commands
 from nycti.discord.nicknames import register_nickname_commands
 from nycti.discord.reminders import register_reminder_commands
-from nycti.discord.rss import register_rss_commands
 from nycti.discord.testing import register_testing_commands
 
 
@@ -23,5 +22,4 @@ def register_bot_commands(bot: Any, *, guild: Any = None) -> None:
     register_memory_commands(bot, guild=guild)
     register_logs_command(bot, guild=guild)
     register_channel_commands(bot, guild=guild)
-    register_rss_commands(bot, guild=guild)
     register_nickname_commands(bot, guild=guild)